requirements.txt
http://download.pytorch.org/whl/cu80/torch-0.1.11.post5-cp35-cp35m-linux_x86_64.whl
numpy
Pillow
scipy
torchvision
h5py
tqdm
